fix up table

+++ back-end/src/elf_section_header_builder.cpp
@@ -34,7 +34,7 @@ lang_status_t add_section_name(buffer_t* tab, const char* name, size_t* offset)
 
 //——————————————————————————————————————————————————————————————————————————————
 
-void build_text_shdr(Elf64_Shdr* shdr, size_t code_size, size_t name_offset)
+void build_text_shdr(Elf64_Shdr* shdr, size_t code_size, uint32_t name_offset)
 {
     ASSERT(shdr);
 
@@ -84,7 +84,7 @@ void build_text_shdr(Elf64_Shdr* shdr, size_t code_size, size_t name_offset)
 
 //——————————————————————————————————————————————————————————————————————————————
 
-void build_shstrtab_shdr(Elf64_Shdr* shdr, size_t shstrtab_size, size_t name_offset, size_t table_offset)
+void build_shstrtab_shdr(Elf64_Shdr* shdr, size_t shstrtab_size, uint32_t name_offset, size_t table_offset)
 {
     ASSERT(shdr);
